Fixed minor typos.
Usage of "e.g." is always followed by a comma and is preceded with a comma or placed in parentheses. It should not start a sentence as it introduces an apposition in a sentence.
Checked consistency of usage for no periods at the end of bullet points, semicolons with "run" before commands, and the oxford comma placement.