Do not restart Docksal services on Docker restart, let fin do it instead.
When Docker for Mac starts it automatically restarts these services. However this means that not the whole system might be started, e.g. resolver or something else might be not setup and there is no way for fin to know it. Preventing auto-restart of the services upon Docker restart will make fin to use system start.