Enable creation of default commands in subfolders
As a project manager / lead developer, I would like to give my dev team a group of commands related to behat, but still allow them to type `fin behat` as a basic command. In order to do this, I need to be able to create a default command that will be executed when the name of the folder is typed. For example, I might have the following commands:

| Command | Runs something like ... |
| --- | --- |
| `fin behat` | `vendor/bin/behat "$@"` |
| `fin behat/info` | `vendor/bin/behat -d i --no-colors > available-tests.txt` |

To do this, I would make a folder structure as follows:

```
.docksal
  commands
    behat
      .default
      info
```